Extract the sync service to a separate module
