+++ bin/full-node/main.rs
@@ -17,33 +17,23 @@
 
 #![recursion_limit = "1024"]
 
-use atomic::Atomic;
-use futures::{
-    channel::{mpsc, oneshot},
-    lock::Mutex,
-    prelude::*,
-};
+use futures::prelude::*;
 use std::{
     borrow::Cow,
-    collections::BTreeMap,
     convert::TryFrom as _,
     fs,
-    net::{SocketAddr, ToSocketAddrs as _},
-    num::{NonZeroU32, NonZeroU64},
     path::PathBuf,
-    pin::Pin,
-    sync::{atomic::Ordering, Arc},
-    thread,
     time::Duration,
 };
 use structopt::StructOpt as _;
 use substrate_lite::{
-    chain::{self, sync::full_optimistic},
-    chain_spec, header, network,
-    network::{connection, multiaddr, peer_id::PeerId, protocol, with_buffers},
+    chain,
+    chain_spec,
+    network::{connection, multiaddr, peer_id::PeerId},
 };
 
 mod network_service;
+mod sync_service;
 
 fn main() {
     env_logger::init();
@@ -137,7 +127,7 @@ async fn async_main() {
     // Load the information about the chain from the database, or build the information of the
     // genesis block.
     // TODO:
-    let chain_information = /*match local_storage.chain_information() {
+    let chain_information_config = /*match local_storage.chain_information() {
         Ok(Some(i)) => i,
         Err(database::local_storage_light::AccessError::StorageAccess(err)) => return Err(err),
         // TODO: log why storage access failed?
@@ -164,9 +154,6 @@ async fn async_main() {
         substrate_lite::metadata::decode(&metadata).unwrap()
     );*/
 
-    let (mut to_sync_tx, to_sync_rx) = mpsc::channel(64);
-    let (to_db_save_tx, mut to_db_save_rx) = mpsc::channel(16);
-
     let network_service = network_service::NetworkService::new(network_service::Config {
         listen_addresses: Vec::new(),
         bootstrap_nodes: {
@@ -191,30 +178,15 @@ async fn async_main() {
     .await
     .unwrap();
 
-    let sync_state = Arc::new(Mutex::new(SyncState {
-        best_block_hash: [0; 32],      // TODO:
-        best_block_number: 0,          // TODO:
-        finalized_block_hash: [0; 32], // TODO:
-        finalized_block_number: 0,     // TODO:
-    }));
-
-    threads_pool.spawn_ok(
-        start_sync(
-            &chain_spec,
-            chain_information,
-            sync_state.clone(),
-            to_sync_rx,
-            network_service.clone(),
-            to_db_save_tx,
-        )
-        .await,
-    );
-
-    threads_pool.spawn_ok(async move {
-        while let Some(info) = to_db_save_rx.next().await {
-            // TODO:
-        }
-    });
+    let sync_service = sync_service::SyncService::new(sync_service::Config {
+        tasks_executor: {
+            let threads_pool = threads_pool.clone();
+            Box::new(move |task| threads_pool.spawn_ok(task))
+        },
+        chain_spec: &chain_spec,
+        chain_information_config,
+    })
+    .await;
 
     /*let mut telemetry = {
         let endpoints = chain_spec
@@ -249,7 +221,7 @@ async fn async_main() {
                     // We end the informant line with a `\r` so that it overwrites itself every time.
                     // If any other line gets printed, it will overwrite the informant, and the
                     // informant will then print itself below, which is a fine behaviour.
-                    let sync_state = sync_state.lock().await.clone();
+                    let sync_state = sync_service.sync_state().await;
                     eprint!("{}\r", substrate_lite::informant::InformantLine {
                         enable_colors: match cli_options.color {
                             ColorChoice::Always => true,
@@ -275,7 +247,20 @@ async fn async_main() {
             network_message = network_service.next_event().fuse() => {
                 match network_message {
                     network_service::Event::Connected(peer_id) => {
-                        to_sync_tx.send(ToSync::NewPeer(peer_id)).await.unwrap();
+                        sync_service.add_source(peer_id).await;
+                    }
+                }
+            }
+
+            sync_message = sync_service.next_event().fuse() => {
+                match sync_message {
+                    sync_service::Event::BlocksRequest { target, request } => {
+                        let block_request = network_service.blocks_request(
+                            target,
+                            request
+                        );
+
+                        todo!()
                     }
                 }
             }
@@ -319,212 +304,4 @@ async fn async_main() {
             },
         }
     }
-}
-
-async fn start_sync(
-    chain_spec: &chain_spec::ChainSpec,
-    chain_information_config: chain::chain_information::ChainInformationConfig,
-    sync_state: Arc<Mutex<SyncState>>,
-    mut to_sync: mpsc::Receiver<ToSync>,
-    network: Arc<network_service::NetworkService>,
-    mut to_db_save_tx: mpsc::Sender<chain::chain_information::ChainInformation>,
-) -> impl Future<Output = ()> {
-    let mut sync =
-        full_optimistic::OptimisticFullSync::<_, network::PeerId>::new(full_optimistic::Config {
-            chain_information_config,
-            sources_capacity: 32,
-            blocks_capacity: {
-                // This is the maximum number of blocks between two consecutive justifications.
-                1024
-            },
-            source_selection_randomness_seed: rand::random(),
-            blocks_request_granularity: NonZeroU32::new(128).unwrap(),
-            download_ahead_blocks: {
-                // Assuming a verification speed of 1k blocks/sec and a 95% latency of one second,
-                // the number of blocks to download ahead of time in order to not block is 1000.
-                1024
-            },
-        });
-
-    let mut finalized_block_storage = BTreeMap::<Vec<u8>, Vec<u8>>::new();
-    // TODO: doesn't necessarily match chain_information; pass this as part of the params of `start_sync` instead
-    for (key, value) in chain_spec.genesis_storage() {
-        finalized_block_storage.insert(key.to_owned(), value.to_owned());
-    }
-
-    async move {
-        let mut peers_source_id_map = hashbrown::HashMap::<_, _, fnv::FnvBuildHasher>::default();
-        let mut block_requests_finished = stream::FuturesUnordered::new();
-
-        loop {
-            // Verify blocks that have been fetched from queries.
-            let mut process = sync.process_one();
-            loop {
-                match process {
-                    full_optimistic::ProcessOne::Idle { sync: s } => {
-                        sync = s;
-                        break;
-                    }
-                    full_optimistic::ProcessOne::Finished {
-                        sync: s,
-                        finalized_blocks,
-                    } => {
-                        process = s.process_one();
-
-                        if let Some(last_finalized) = finalized_blocks.last() {
-                            let mut lock = sync_state.lock().await;
-                            lock.finalized_block_hash = last_finalized.header.hash();
-                            lock.finalized_block_number = last_finalized.header.number;
-                        }
-
-                        for block in finalized_blocks {
-                            for (key, value) in block.storage_top_trie_changes {
-                                if let Some(value) = value {
-                                    finalized_block_storage.insert(key, value);
-                                } else {
-                                    let _was_there = finalized_block_storage.remove(&key);
-                                    // TODO: if a block inserts a new value, then removes it in the next block, the key will remain in `finalized_block_storage`; either solve this or document this
-                                    // assert!(_was_there.is_some());
-                                }
-                            }
-                        }
-                    }
-
-                    full_optimistic::ProcessOne::InProgress {
-                        current_best_hash,
-                        current_best_number,
-                        resume,
-                    } => {
-                        // Processing has made a step forward.
-                        // There is nothing to do, but this is used to update to best block
-                        // shown on the informant.
-                        let mut lock = sync_state.lock().await;
-                        lock.best_block_hash = current_best_hash;
-                        lock.best_block_number = current_best_number;
-                        drop(lock);
-
-                        process = resume.resume();
-                    }
-
-                    full_optimistic::ProcessOne::FinalizedStorageGet(req) => {
-                        let value = finalized_block_storage
-                            .get(&req.key_as_vec())
-                            .map(|v| &v[..]);
-                        process = req.inject_value(value);
-                    }
-                    full_optimistic::ProcessOne::FinalizedStorageNextKey(req) => {
-                        // TODO: to_vec() :-/
-                        let req_key = req.key().to_vec();
-                        // TODO: to_vec() :-/
-                        let next_key = finalized_block_storage
-                            .range(req.key().to_vec()..)
-                            .skip_while(move |(k, _)| &k[..] <= &req_key[..])
-                            .next()
-                            .map(|(k, _)| k);
-                        process = req.inject_key(next_key);
-                    }
-                    full_optimistic::ProcessOne::FinalizedStoragePrefixKeys(req) => {
-                        // TODO: to_vec() :-/
-                        let prefix = req.prefix().to_vec();
-                        // TODO: to_vec() :-/
-                        let keys = finalized_block_storage
-                            .range(req.prefix().to_vec()..)
-                            .take_while(|(k, _)| k.starts_with(&prefix))
-                            .map(|(k, _)| k);
-                        process = req.inject_keys(keys);
-                    }
-                }
-            }
-
-            // Update the current best block, used for CLI-related purposes.
-            {
-                let mut lock = sync_state.lock().await;
-                lock.best_block_hash = sync.best_block_hash();
-                lock.best_block_number = sync.best_block_number();
-            }
-
-            // Start requests that need to be started.
-            // Note that this is done after calling `process_one`, as the processing of pending
-            // blocks can result in new requests but not the contrary.
-            while let Some(action) = sync.next_request_action() {
-                match action {
-                    full_optimistic::RequestAction::Start {
-                        start,
-                        block_height,
-                        source,
-                        num_blocks,
-                        ..
-                    } => {
-                        let block_request = network.blocks_request(
-                            source.clone(),
-                            protocol::BlocksRequestConfig {
-                                start: protocol::BlocksRequestConfigStart::Number(block_height),
-                                desired_count: num_blocks,
-                                direction: protocol::BlocksRequestDirection::Ascending,
-                                fields: protocol::BlocksRequestFields {
-                                    header: true,
-                                    body: true,
-                                    justification: true,
-                                },
-                            },
-                        );
-                        let (block_request, abort) = future::abortable(block_request);
-                        let request_id = start.start(abort);
-                        block_requests_finished.push(block_request.map(move |r| (request_id, r)));
-                    }
-                    full_optimistic::RequestAction::Cancel { user_data, .. } => {
-                        user_data.abort();
-                    }
-                }
-            }
-
-            futures::select! {
-                message = to_sync.next() => {
-                    let message = match message {
-                        Some(m) => m,
-                        None => return,
-                    };
-
-                    match message {
-                        ToSync::NewPeer(peer_id) => {
-                            let id = sync.add_source(peer_id.clone());
-                            peers_source_id_map.insert(peer_id.clone(), id);
-                        },
-                        ToSync::PeerDisconnected(peer_id) => {
-                            let id = peers_source_id_map.remove(&peer_id).unwrap();
-                            let (_, rq_list) = sync.remove_source(id);
-                            for (_, rq) in rq_list {
-                                rq.abort();
-                            }
-                        },
-                    }
-                },
-
-                (request_id, result) = block_requests_finished.select_next_some() => {
-                    // `result` is an error if the block request got cancelled by the sync state
-                    // machine.
-                    if let Ok(result) = result {
-                        let _ = sync.finish_request(request_id, result.map(|v| v.into_iter().map(|block| full_optimistic::RequestSuccessBlock {
-                            scale_encoded_header: block.header.unwrap(), // TODO: don't unwrap
-                            scale_encoded_extrinsics: block.body.unwrap(), // TODO: don't unwrap
-                            scale_encoded_justification: block.justification,
-                        })).map_err(|()| full_optimistic::RequestFail::BlocksUnavailable));
-                    }
-                },
-            }
-        }
-    }
-}
-
-enum ToSync {
-    NewPeer(network::PeerId),
-    PeerDisconnected(network::PeerId),
-}
-
-#[derive(Debug, Clone)]
-struct SyncState {
-    best_block_number: u64,
-    best_block_hash: [u8; 32],
-    finalized_block_number: u64,
-    finalized_block_hash: [u8; 32],
 }
